+++ build.rs
@@ -3,7 +3,7 @@ use self_update::backends::github::Update;
 use std::{env, fs::File, path::PathBuf};
 
 fn main() -> Result<()> {
-	let out_path = PathBuf::from(env::var("OUT_DIR")?).join("Argon.rbxm");
+	let out_path = PathBuf::from(env::var("OUT_DIR")?).join("Lemonade.rbxm");
 
 	if !cfg!(feature = "plugin") {
 		File::create(out_path)?;
@@ -21,14 +21,14 @@ fn main() -> Result<()> {
 	builder
 		.repo_owner("LupaHQ")
 		.repo_name("argon-roblox")
-		.bin_name("Argon.rbxm")
+		.bin_name("Lemonade.rbxm")
 		.bin_install_path(out_path)
 		.target("");
 
 	builder
 		.build()?
 		.download()
-		.context("Failed to download Argon plugin from GitHub!")?;
+		.context("Failed to download Lemonade plugin from GitHub!")?;
 
 	Ok(())
 }