+++ src/util.rs
@@ -31,7 +31,7 @@ pub fn get_username() -> String {
 }
 
 pub fn get_plugin_path() -> Result<PathBuf> {
-	Ok(RobloxStudio::locate()?.plugins_path().join("Argon.rbxm"))
+	Ok(RobloxStudio::locate()?.plugins_path().join("Lemonade.rbxm"))
 }
 
 /// Checks if the given `class` is a service